+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/recovery_lineage/retry_stalls/retry_backoff.rs
@@ -1,0 +1,145 @@
+use crate::orchestrator::tests::operator::status::running_lanes::{
+	self, OffsetDateTime, StateStore, fs, orchestrator, process, state,
+};
+
+#[test]
+fn operator_status_snapshot_reports_retry_backoff_from_worktree_marker() {
+	for (retry_kind, expected_wait_reason) in
+		[("failure", "failure_retry"), ("git_lock_contention", "git_lock_contention")]
+	{
+		let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+		let state_store = StateStore::open_in_memory().expect("state store should open");
+		let issue = running_lanes::sample_issue("Todo", &[]);
+		let worktree_path = config.worktree_root().join("PUB-101");
+
+		state_store
+			.record_run_attempt("run-1", &issue.id, 1, "failed")
+			.expect("run attempt should record");
+		state_store
+			.upsert_worktree(
+				"pubfi",
+				&issue.id,
+				"x/pubfi-pub-101",
+				&worktree_path.display().to_string(),
+			)
+			.expect("worktree should record");
+
+		state::write_run_retry_schedule(
+			&worktree_path,
+			"run-1",
+			1,
+			retry_kind,
+			OffsetDateTime::now_utc().unix_timestamp() + 60,
+		)
+		.expect("retry schedule marker should write");
+
+		let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+			.expect("snapshot should build");
+		let run = snapshot.recent_runs.first().expect("recent run should exist");
+
+		assert_eq!(run.phase, "retry_backoff");
+		assert_eq!(run.wait_reason.as_deref(), Some(expected_wait_reason));
+		assert_eq!(run.retry_kind.as_deref(), Some(retry_kind));
+		assert!(run.next_retry_at.is_some());
+		assert_eq!(snapshot.projects[0].waiting_lane_count, 1);
+		assert_eq!(snapshot.projects[0].connector_state, "backoff");
+	}
+}
+
+#[test]
+fn operator_status_snapshot_keeps_continuation_retry_from_orphaning_live_marker_worktree() {
+	let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = running_lanes::sample_issue("Todo", &[]);
+	let worktree_path = config.worktree_root().join("PUB-101");
+
+	state_store
+		.record_run_attempt("run-1", &issue.id, 1, "continuation_pending")
+		.expect("run attempt should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-101",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
+	state::write_run_activity_marker_for_process(&worktree_path, "run-1", 1, process::id())
+		.expect("live marker should write");
+	state::write_run_retry_schedule(
+		&worktree_path,
+		"run-1",
+		1,
+		"continuation",
+		OffsetDateTime::now_utc().unix_timestamp() + 60,
+	)
+	.expect("continuation retry schedule marker should write");
+
+	let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+		.expect("snapshot should build");
+	let run = snapshot.recent_runs.first().expect("recent run should exist");
+	let rendered = orchestrator::render_operator_status(&snapshot);
+
+	assert!(snapshot.current_lanes.is_empty());
+	assert_eq!(run.status, "continuation_pending");
+	assert_eq!(run.phase, "retry_backoff");
+	assert_eq!(run.wait_reason.as_deref(), Some("continuation_retry"));
+	assert_eq!(run.retry_kind.as_deref(), Some("continuation"));
+	assert_eq!(run.ownership_state, "continuation_pending");
+	assert_eq!(run.liveness_state, "process_alive");
+	assert_eq!(run.lane_control_next_action, "wait_for_continuation_reentry");
+	assert_eq!(snapshot.worktrees[0].ownership, "continuation_pending");
+	assert_eq!(snapshot.projects[0].retained_worktree_count, 0);
+	assert_eq!(snapshot.projects[0].waiting_lane_count, 1);
+	assert!(rendered.contains("Recovery worktrees: 0"));
+	assert!(rendered.contains("- none (owned worktrees are shown in their lane sections above)"));
+	assert!(!rendered.contains("role: orphaned_live_thread"));
+}
+
+#[test]
+fn operator_status_snapshot_ignores_retry_schedule_on_running_attempt() {
+	let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = running_lanes::sample_issue("Todo", &[]);
+	let worktree_path = config.worktree_root().join("PUB-101");
+
+	state_store
+		.record_run_attempt("run-1", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", &issue.id, "run-1", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-101",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
+	state::write_run_activity_marker_for_process(&worktree_path, "run-1", 1, process::id())
+		.expect("run marker should write");
+	state::write_run_retry_schedule(
+		&worktree_path,
+		"run-1",
+		1,
+		"failure",
+		OffsetDateTime::now_utc().unix_timestamp() + 60,
+	)
+	.expect("stale retry schedule marker should write");
+
+	let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+		.expect("snapshot should build");
+	let run = snapshot.current_lanes.first().expect("current lane should exist");
+
+	assert_eq!(run.phase, "executing");
+	assert_eq!(run.wait_reason, None);
+	assert_eq!(run.retry_kind, None);
+	assert_eq!(run.next_retry_at, None);
+	assert_eq!(snapshot.projects[0].waiting_lane_count, 0);
+	assert_eq!(snapshot.projects[0].connector_state, "ok");
+}
